int instead of ssize_t (to ease sbpl standalone build)
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@7759 f5854215-dd47-0410-b2c4-cdd35faa7885

+++ motion_planning/sbpl_util/src/sbpl_util.cpp
@@ -163,7 +163,7 @@ namespace ompl {
 	 double * actual_time_wall_sec,
 	 double * actual_time_user_sec,
 	 double * actual_time_system_sec,
-	 ssize_t * number_of_expands,
+	 int * number_of_expands,
 	 int * solution_cost,
 	 double * solution_epsilon,
 	 vector<int>* solution_stateIDs_V) throw(no_planner_selected)
